# api_documentation/Core.swagger_collection.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0",
    "title": "Deep Lynx",
...
    "contact": {}
  },
  "host": "example.com",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/containers/{container-id}": {
      "put": {
        "description": "Updates the container. This will fail if a container already exists with the proposed updated name.",
        "summary": "Update",
        "tags": [
          "Containers"
        ],
        "operationId": "Update",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "container-id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "Content-Type",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/UpdateRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      },
      "get": {
        "description": "Fetch container by ID.",
        "summary": "Fetch by ID",
        "tags": [
          "Containers"
        ],
        "operationId": "FetchbyID",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "container-id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      },
      "delete": {
        "description": "Archives a Container. This is preferred over deletion as deletion has a cascading effect on the deleted type's keys, relationships, and relationship keys. When in doubt, archive over delete. We'd rather have tombstones than cremating the type.",
        "summary": "Archive",
        "tags": [
          "Containers"
        ],
        "operationId": "Archive",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "container-id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
    "/containers/{container-id}/metatypes": {
      "post": {
        "description": "Create a new meta type. Pass in an array for bulk creation.",
        "summary": "New Type",
        "tags": [
          "Metatypes"
        ],
        "operationId": "NewType",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "container-id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "Content-Type",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/NewTypeRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      },
      "get": {
        "description": "List all types that current container has access to. MUST include the limit/offset query parameters as of 3/26/2020",
        "summary": "List Types",
        "tags": [
          "Metatypes"
        ],
        "operationId": "ListTypes",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": true,
            "type": "integer",
            "format": "int32",
            "description": ""
          },
          {
            "name": "offset",
            "in": "query",
            "required": true,
            "type": "integer",
            "format": "int32",
            "description": ""
          },
          {
            "name": "container-id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
    "/containers/{container-id}/metatypes/{type-id}": {
      "get": {
        "description": "Retrieves a single Type.",
        "summary": "Get Type",
        "tags": [
          "Metatypes"
        ],
        "operationId": "GetType",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "container-id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "type-id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      },
      "put": {
        "description": "Update a single Metatype in storage. Will fail if the updated name has already been taken.",
        "summary": "Update Type",
        "tags": [
          "Metatypes"
        ],
        "operationId": "UpdateType",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "container-id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "type-id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "Content-Type",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/UpdateTypeRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      },
      "delete": {
        "description": "Archives the current meta type. This is preferred over deletion as deletion has a cascading effect on the deleted type's keys, relationships, and relationship keys. When in doubt, archive over delete. We'd rather have tombstones than cremating the type.",
        "summary": "Archive Type",
        "tags": [
          "Metatypes"
        ],
        "operationId": "ArchiveType",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "container-id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "type-id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
...
    "/containers/{container-id}/metatype_relationships": {
      "post": {
        "description": "Create a new Type Relationship. Describes the connection that could exist between two types and acts as a vehicle for relationship specific keys.\n\nPass in an array for bulk creation.",
        "summary": "New Type Relationship",
        "tags": [
          "Metatype Relationships"
        ],
        "operationId": "NewTypeRelationship",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "container-id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "Content-Type",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/NewTypeRelationshipRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      },
      "get": {
        "description": "Create a new Type relationship pair. Describes the connection between two types and acts as a vehicle for relationship specific keys.",
        "summary": "List Container Type Relationships",
        "tags": [
          "Metatype Relationships"
        ],
        "operationId": "ListContainerTypeRelationships",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "offset",
            "in": "query",
            "required": true,
            "type": "integer",
            "format": "int32",
            "description": ""
          },
          {
            "name": "limit",
            "in": "query",
            "required": true,
            "type": "integer",
            "format": "int32",
            "description": ""
          },
          {
            "name": "container-id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
    "/containers/{container-id}/metatype_relationships/{relationship-id}": {
      "get": {
        "description": "Retrieve a single Type Relationship.",
        "summary": "Get Type Relationship",
        "tags": [
          "Metatype Relationships"
        ],
        "operationId": "GetTypeRelationship",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "container-id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "relationship-id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      },
      "put": {
        "summary": "Update Type Relationship",
        "tags": [
          "Metatype Relationships"
        ],
        "operationId": "UpdateTypeRelationship",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "container-id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "relationship-id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "Content-Type",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/UpdateTypeRelationshipRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      },
      "delete": {
        "description": "Archive the type relationship key. This does not affect any instances of the parent relationship type.",
        "summary": "Archive Type Relationship",
        "tags": [
          "Metatype Relationships"
        ],
        "operationId": "ArchiveTypeRelationship",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "container-id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "relationship-id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "UpdateRequest": {
      "title": "UpdateRequest",
      "example": {
        "description": "Final Test Description"
      },
      "type": "object",
      "properties": {
        "description": {
          "type": "string"
        }
      },
      "required": [
        "description"
      ]
    },
...
    "NewTypeRequest": {
      "title": "NewTypeRequest",
      "example": {
        "name": "Equipment",
        "description": "Physical piece of equipement"
      },
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        }
      },
      "required": [
        "name",
        "description"
      ]
    },
    "UpdateTypeRequest": {
      "title": "UpdateTypeRequest",
      "example": {
        "name": "Updated Metatype Name",
        "description": "Final Description"
      },
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        }
      },
      "required": [
        "name",
        "description"
      ]
    },
...
    "NewTypeRelationshipRequest": {
      "title": "NewTypeRelationshipRequest",
      "example": {
        "name": "Relationhip Test",
        "description": "string"
      },
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        }
      },
      "required": [
        "name",
        "description"
      ]
    },
    "UpdateTypeRelationshipRequest": {
      "title": "UpdateTypeRelationshipRequest",
      "example": {
        "name": "Update Test",
        "description": "Final Test Description"
      },
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        }
      },
      "required": [
        "name",
        "description"
      ]
    },
...
  },
...
}